GameBoardTest milestone
Here's where it was when Ian and I broke for the night around 12:30 this
morning.

I updated the GameBoardTest class. It's now linked to a deck so it is
essentially a prototype for our draw pile. It's a little wonky if you
don't drag the card to the right place, but that's because the default
action I envision in the situation, adding it to the hand, doesn't exist
yet. Also updated Card and House and some of the pictures for consistent
so the GUI can grab the pictures.