Updated GameBoardTest
You can virtually play your own game at this point. It's just that the
other players are limited to what they draw or you share a common hand.

Lots of image changes to go with the current state of the GameBoardTest

Updated Game and Player to their most recent as well...