Take 2 on the most recent changes
Updated GameBoardTest - You can basically play the game, but it's not
linked to anything yet. I've put output statements for most events which
should make it straightforward to pass messages to the game class at the
appropriate time. Background graphic still needs to be updated with game
area outlines and feedback.